Allow plot plugin to have tools
