+++ skimage/viewer/canvastools/painttool.py
@@ -14,8 +14,8 @@ class PaintTool(CanvasToolBase):
 
     Parameters
     ----------
-    viewer : :class:`skimage.viewer.Viewer`
-        Skimage viewer object.
+    manager : Viewer or PlotPlugin.
+        Skimage viewer or plot plugin object.
     overlay_shape : shape tuple
         2D shape tuple used to initialize overlay image.
     alpha : float (between [0, 1])
@@ -38,10 +38,10 @@ class PaintTool(CanvasToolBase):
     label : int
         Current paint color.
     """
-    def __init__(self, viewer, overlay_shape, radius=5, alpha=0.3,
+    def __init__(self, manager, overlay_shape, radius=5, alpha=0.3,
                  on_move=None, on_release=None, on_enter=None,
                  rect_props=None):
-        super(PaintTool, self).__init__(viewer, on_move=on_move,
+        super(PaintTool, self).__init__(manager, on_move=on_move,
                                         on_enter=on_enter,
                                         on_release=on_release)
 
@@ -63,7 +63,7 @@ class PaintTool(CanvasToolBase):
 
         # Note that the order is important: Redraw cursor *after* overlay
         self.artists = [self._overlay_plot, self._cursor]
-        viewer.add_tool(self)
+        self.manager.add_tool(self)
 
     @property
     def label(self):
@@ -136,7 +136,7 @@ class PaintTool(CanvasToolBase):
         self.callback_on_release(self.geometry)
 
     def on_move(self, event):
-        if not self.viewer.ax.in_axes(event):
+        if not self.ax.in_axes(event):
             self._cursor.set_visible(False)
             self.redraw() # make sure cursor is not visible
             return
